refactor(card): DRY getTemperatureColor
